requirements.txt
torch>=2.1.0
transformers>=4.36.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
optimum>=1.16.0
peft>=0.7.0
omegaconf>=2.3.0
pyyaml>=6.0
datasets>=2.16.0
pandas>=2.1.0
scipy>=1.11.0
numpy>=1.26.0
scikit-learn>=1.3.0
tqdm>=4.66.0
ruff>=0.1.0
black>=23.12.0
mypy>=1.8.0
jupyter>=1.0.0
pytest>=7.0.0
matplotlib>=3.8.0
seaborn